requirements.txt.txt
streamlit
opencv-python-headless
numpy
matplotlib
scipy
Pillow